gens_core/vdp/vdp_rend_m5.cpp: Ported GET_PATTERN_INFO to C++.
This function gets the pattern information (tile number, palette, etc.)
for a scroll plane at the specified coordinates.

+++ src/gens/gens_core/vdp/vdp_io.h
@@ -167,8 +167,8 @@ extern VDP_Flags_t VDP_Flags;
 extern int Zero_Length_DMA;
 
 // VDP address pointers.
-extern uint8_t *ScrA_Addr;
-extern uint8_t *ScrB_Addr;
+extern uint16_t *ScrA_Addr;
+extern uint16_t *ScrB_Addr;
 extern uint8_t *Win_Addr;
 extern uint8_t *Spr_Addr;
 extern uint16_t *H_Scroll_Addr;
